Make Lib Build in Non-AMD64 Architectures
This commit adds a mechanism for the `lib/Makefile` to determine the arch for
which it is being invoked to build. It then uses this to decide whether or not
to compile the amd64 asm file.

This also includes minor / temporary changes to the source to make it build
when `TARGET_ATTRIBUTE("bmi2")` is not available.

Note that while `make lib` works with this, most other invocations still fail,
because other makefiles in the zstd project do their own source file globbing
(and even building!!). This will need to be fixed separately.