[licensing] Remove Examples Section from Bottom of GPL License File
As I understand it, the `How to Apply These Terms to Your New Programs`
section is not actually part of the license and can be removed.

The motivation to remove it is the concern that the example language therein
might be confusing for `grep` or for people trying to understand Zstd's
licensing status.